Refactor: sign_up 폴더 파일 이동 완료
